implement: Unit test for line_append

+++ lines.c
@@ -14,7 +14,7 @@ void lines_append_last(struct lines *p, struct line *line) {
 }
 
 void lines_free(struct lines *p) {
-	for (int i = 0; i < p->arr->size; i++)
+	for (size_t i = 0; i < p->arr->size; i++)
 		line_free(p->arr->data[i]);
 	MemFree(p->arr);
 	MemFree(p);
